refactor(rust-sys): relocate examples
