+++ examples/wasmedge-sys/async_host_func.rs
