+++ examples/wasmedge-sys/async_run_func.rs
