+++ examples/wasmedge-sys/async_run_registered_func.rs
