+++ examples/wasmedge-sys/context_data_to_host_func.rs
