+++ examples/wasmedge-sys/data/fibonacci.wat
