+++ examples/wasmedge-sys/data/funcs.wat
