+++ examples/wasmedge-sys/data/import.wat
