+++ examples/wasmedge-sys/data/module1.wat
