+++ examples/wasmedge-sys/data/module2.wat
