+++ examples/wasmedge-sys/data/test.wasm
