+++ examples/wasmedge-sys/data/test.wat
