+++ examples/wasmedge-sys/executor_register_active_module.rs
