+++ examples/wasmedge-sys/executor_register_named_module.rs
