+++ examples/wasmedge-sys/fibonacci.rs
