+++ examples/wasmedge-sys/func_call_across_different_modes.rs
