+++ examples/wasmedge-sys/hostfunc.rs
