+++ examples/wasmedge-sys/hostfunc2.rs
