+++ examples/wasmedge-sys/mdbook_example_module_instance.rs
