+++ examples/wasmedge-sys/store_get_named_module.rs
