+++ src/instance.rs
@@ -267,7 +267,7 @@ mod tests {
         // add a wasm module from a file
         let file = std::env::current_dir()
             .unwrap()
-            .join("crates/wasmedge-sys/examples/data/fibonacci.wat");
+            .join("examples/wasmedge-sys/data/fibonacci.wat");
         let result = Module::from_file(Some(&config), file);
         assert!(result.is_ok());
         let module = result.unwrap();
